pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.2",
    "anthropic>=0.75.0",
    "asyncpg>=0.31.0",
    "azure-identity>=1.25.1",
    "azure-mgmt-resource>=24.0.0",
    "bcrypt>=5.0.0",
    "boto3>=1.41.5",
    "botocore>=1.41.5",
    "celery>=5.5.3",
    "docx>=0.2.4",
    "email-validator>=2.3.0",
    "fastapi>=0.122.0",
    "gitpython>=3.1.45",
    "google-auth>=2.43.0",
    "google-cloud-compute>=1.40.0",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "hiredis>=3.3.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "python-engineio>=4.12.3",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "python-socketio>=5.15.0",
    "python-terraform>=0.10.1",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "shortuuid>=1.0.13",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
